Cast SimpleStorage generated byte count to int
Calculations of the CapacityBytes for the SimpleStorage Device were
causing it to be treated as a float, resulting in the returned size
being returned as 549755813888.0. The presence of the decimal point
causes some clients to treat the value as a floating point value, even
though it is a full integer.

Since the spec declares this property as being an integer, this updates
the code to cast the calculated size to an int to make sure it is
returned as one via the API call.

Signed-off-by: Sean McGinnis <sean.mcginnis@gmail.com>